add GetRegionName
The region code returned from City queries is not useful by itself.  To
turn this into a human readable region, we can either query the
region.csv file by hand or call the GeoIP_region_name_by_code which
contains a generated lookup table to do the mapping for us.