Return netmask in City lookups
While we could retrieve this information for successful lookups with the
prior changes, we could not do so if the lookups failed. Being able to
know this when the lookup failed is useful for knowing how large the gap
in the database is so that we can skip over it to the next.